Extract collection parsers to separate file
